backend/requirements.txt
fastapi
uvicorn
websockets
crewai
langchain-google-genai
python-dotenv
mesa